+++ src/Symfony/Component/Security/Core/Tests/Fixtures/Enum/IntBackedEnum.php
@@ -1,0 +1,20 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Symfony\Component\Security\Core\Tests\Fixtures\Enum;
+
+enum IntBackedEnum: int
+{
+    case FOO = 0;
+    case BAR = 1;
+}